move final tests from LedgerSpec
