+++ src/test/scala/io/iohk/ethereum/ledger/BlockPreparatorSpec.scala
@@ -328,4 +328,76 @@ class BlockPreparatorSpec extends AnyWordSpec with Matchers with ScalaCheckPrope
     result match { case (_, executedTxs) => executedTxs shouldBe Seq.empty }
   }
 
+  // migrated from old LedgerSpec
+  "properly assign stateRootHash before byzantium block (exclusive)" in new TestSetup {
+
+    val tx: Transaction = defaultTx.copy(
+      gasPrice = defaultGasPrice,
+      gasLimit = defaultGasLimit,
+      receivingAddress = None,
+      payload = ByteString.empty
+    )
+    val stx: SignedTransactionWithSender = SignedTransaction.sign(tx, originKeyPair, Some(blockchainConfig.chainId))
+    val header: BlockHeader =
+      defaultBlockHeader.copy(number = blockchainConfig.forkBlockNumbers.byzantiumBlockNumber - 1)
+
+    val result: Either[BlockExecutionError.TxsExecutionError, BlockResult] =
+      consensus.blockPreparator.executeTransactions(Seq(stx.tx), initialWorld, header)
+
+    result shouldBe a[Right[_, BlockResult]]
+    result.map { br =>
+      br.receipts.last.postTransactionStateHash shouldBe a[HashOutcome]
+    }
+  }
+
+  "properly assign stateRootHash after byzantium block (inclusive) if operation is a success" in new TestSetup {
+
+    val tx: Transaction = defaultTx.copy(
+      gasPrice = defaultGasPrice,
+      gasLimit = defaultGasLimit,
+      receivingAddress = None,
+      payload = ByteString.empty
+    )
+    val stx: SignedTransaction = SignedTransaction.sign(tx, originKeyPair, Some(blockchainConfig.chainId)).tx
+    val header: BlockHeader =
+      defaultBlockHeader.copy(
+        beneficiary = minerAddress.bytes,
+        number = blockchainConfig.forkBlockNumbers.byzantiumBlockNumber
+      )
+
+    val result: Either[BlockExecutionError.TxsExecutionError, BlockResult] =
+      consensus.blockPreparator.executeTransactions(Seq(stx), initialWorld, header)
+
+    result shouldBe a[Right[_, BlockResult]]
+    result.map(_.receipts.last.postTransactionStateHash shouldBe SuccessOutcome)
+  }
+
+  "properly assign stateRootHash after byzantium block (inclusive) if operation is a failure" in new TestSetup {
+
+    val defaultsLogs = Seq(defaultLog)
+
+    lazy val mockVM =
+      new MockVM(createResult(_, defaultGasLimit, defaultGasLimit, 0, Some(RevertOccurs), bEmpty, defaultsLogs))
+
+    val testConsensus: Consensus = newTestConsensus(vm = mockVM)
+
+    val tx: Transaction = defaultTx.copy(
+      gasPrice = defaultGasLimit,
+      gasLimit = defaultGasLimit,
+      receivingAddress = None,
+      payload = ByteString.empty
+    )
+    val stx: SignedTransactionWithSender = SignedTransaction.sign(tx, originKeyPair, Some(blockchainConfig.chainId))
+    val header: BlockHeader =
+      defaultBlockHeader.copy(
+        beneficiary = minerAddress.bytes,
+        number = blockchainConfig.forkBlockNumbers.byzantiumBlockNumber
+      )
+
+    val result: Either[BlockExecutionError.TxsExecutionError, BlockResult] =
+      testConsensus.blockPreparator.executeTransactions(Seq(stx.tx), initialWorld, header)
+
+    result shouldBe a[Right[_, BlockResult]]
+    result.map(_.receipts.last.postTransactionStateHash shouldBe FailureOutcome)
+  }
 }
